proyecto_Django/requirements.txt
asgiref==3.8.1
backports.zoneinfo==0.2.1
certifi==2024.7.4
charset-normalizer==3.3.2
ci-info==0.3.0
click==8.1.7
configobj==5.0.8
configparser==7.0.0
Django==4.2.14
django-cors-headers==4.4.0
djangorestframework==3.15.2
etelemetry==0.3.1
filelock==3.15.4
fitz==0.0.1.dev2
httplib2==0.22.0
idna==3.7
importlib-resources==6.4.0
isodate==0.6.1
looseversion==1.3.0
lxml==5.2.2
networkx==3.1
nibabel==5.2.1
nipype==1.8.6
numpy==1.24.4
packaging==24.1
pandas==2.0.3
pathlib==1.0.1
pillow==10.4.0
prov==2.0.1
psycopg2-binary==2.9.9
pydot==3.0.1
PyMuPDF==1.24.8
PyMuPDFb==1.24.8
pyparsing==3.1.2
python-dateutil==2.9.0.post0
pytz==2024.1
pyxnat==1.6.2
rdflib==7.0.0
requests==2.32.3
scipy==1.10.1
simplejson==3.19.2
six==1.16.0
sqlparse==0.5.1
traits==6.3.2
typing-extensions==4.12.2
tzdata==2024.1
urllib3==2.2.2
zipp==3.19.2